@@ -8,18 +8,20 @@
     "start": "next start"
   },
   "dependencies": {
-    "autoprefixer": "^10.1.0",
-    "mongoose": "^5.11.11",
-    "next": "10.0.4",
-    "postcss": "^8.2.2",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "tailwindcss": "^2.0.2"
+    "autoprefixer": "^10.4.13",
+    "http-react": "^3.0.1",
+    "mongoose": "^6.9.0",
+    "next": "^13.1.6",
+    "postcss": "^8.4.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tailwindcss": "^3.2.4"
   },
   "devDependencies": {
-    "@types/mongoose": "^5.10.3",
-    "@types/node": "^14.14.20",
-    "@types/react": "^17.0.0",
-    "typescript": "^4.1.3"
+    "@types/mongoose": "^5.11.97",
+    "@types/node": "^18.11.19",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "typescript": "^4.9.5"
   }
 }
